Implemented first command
