--- .gitignore
@@ -1,5 +1,0 @@
-main.go
-vendor
-
-.code
-.idea
