+++ lib/.gitignore
@@ -1,0 +1,5 @@
+main.go
+vendor
+
+.code
+.idea
