+++ lib/docker-compose.yml
@@ -1,0 +1,49 @@
+version: '2'
+services:
+  connect:
+    image: confluentinc/cp-kafka-connect:3.3.0-1
+    ports:
+      - "8083:8083"
+    links:
+      - kafka:kafka
+    environment:
+      - CONNECT_GROUP_ID=connect-cluster
+
+      - CONNECT_KEY_CONVERTER=org.apache.kafka.connect.json.JsonConverter
+      - CONNECT_VALUE_CONVERTER=org.apache.kafka.connect.json.JsonConverter
+      - CONNECT_KEY_CONVERTER_SCHEMAS_ENABLE=false
+      - CONNECT_VALUE_CONVERTER_SCHEMAS_ENABLE=false
+
+      - CONNECT_INTERNAL_KEY_CONVERTER=org.apache.kafka.connect.json.JsonConverter
+      - CONNECT_INTERNAL_VALUE_CONVERTER=org.apache.kafka.connect.json.JsonConverter
+      - CONNECT_INTERNAL_KEY_CONVERTER_SCHEMAS_ENABLE=false
+      - CONNECT_INTERNAL_VALUE_CONVERTER_SCHEMAS_ENABLE=false
+
+      - CONNECT_OFFSET_STORAGE_TOPIC=connect-offsets
+      - CONNECT_CONFIG_STORAGE_TOPIC=connect-configs
+      - CONNECT_STATUS_STORAGE_TOPIC=connect-status
+      - CONNECT_OFFSET_STORAGE_REPLICATION_FACTOR=1
+      - CONNECT_CONFIG_STORAGE_REPLICATION_FACTOR=1
+      - CONNECT_STATUS_STORAGE_REPLICATION_FACTOR=1
+
+      - CONNECT_OFFSET_FLUSH_INTERVAL_MS=10000
+      - CONNECT_REST_ADVERTISED_HOST_NAME=localhost
+
+      - CONNECT_BOOTSTRAP_SERVERS=kafka:9092
+
+      - CONNECT_LOG4J_ROOT_LOGLEVEL=WARN
+
+  zookeeper:
+    image: wurstmeister/zookeeper
+    ports:
+      - "2181:2181"
+
+  kafka:
+    image: wurstmeister/kafka
+    ports:
+      - "9092:9092"
+    environment:
+      KAFKA_PORT: 9092
+      KAFKA_ADVERTISED_PORT: 9092
+      KAFKA_ADVERTISED_HOST_NAME: kafka
+      KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
